requirements.txt
Flask>=2.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0
protobuf>=5.0.0
torch>=1.9.0
numpy>=1.21.0
Pillow>=8.0.0
python-dotenv>=0.19.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
